+++ options.js
@@ -1,5 +1,5 @@
 const eslint = require('eslint')
-const config = require('./dist/config')
+const plugins = require('./dist/js/plugins').plugins
 const join = require('path').join
 const pkg = require('./package.json')
 
@@ -9,11 +9,13 @@ module.exports = {
   bugs         : pkg.bugs.url,
   cmd          : 'lint',
   eslint,
+  cwd          : '',
   eslintConfig : {
-    cache   : true,
+    cache       : true,
     configFile,
-    fix     : false,
-    plugins : config.plugins
+    fix         : true,
+    plugins,
+    useEslintrc : true
   },
   homepage : pkg.homepage,
   tagline  : 'Honesty in small things is not a small thing.',