+++ src/base/env.ts
@@ -1,0 +1,9 @@
+namespace BaseConfiguration {
+
+  export const env = {
+    commonjs: true,
+    es6: true,
+    node: true
+  }
+
+}