+++ src/base/extends.ts
@@ -1,0 +1,5 @@
+namespace BaseConfiguration {
+
+  export const extend = []
+
+}