+++ src/base/globals.ts
@@ -1,0 +1,17 @@
+namespace BaseConfiguration {
+
+  export const globals = {
+    __DEV__ : true,
+    __SERVER__ : true,
+    __ : true,
+    define : true,
+    describe : true,
+    expect : true,
+    require : true,
+    test : true,
+    document : true,
+    navigator : false,
+    window : true
+  }
+
+}