+++ src/output/env.ts
@@ -1,9 +1,0 @@
-export const env = {
-  amd: true,
-  browser: true,
-  commonjs: true,
-  es6: true,
-  jest: true,
-  jquery: true,
-  node: true
-}