+++ src/output/globals.ts
@@ -1,5 +1,0 @@
-export const globals = {
-  document: true,
-  navigator: false,
-  window: true
-}