+++ src/overrides/index.ts
@@ -1,0 +1,3 @@
+import { ts } from './overrides'
+
+export const overrides = [ts]