+++ src/overrides/overrides.ts
@@ -1,0 +1,12 @@
+export const ts = {
+  files: ['**/*.ts'],
+  parser: 'typescript-eslint-parser'
+}
+
+export const md = {
+  // To do
+}
+
+export const jest = {
+  files: ['**/*.test.js']
+}